requirements.txt
streamlit
google-generativeai
pandas
altair
python-dotenv